change: zoomimage-view-sketch and zoomimage-compose-sketch modules upgraded to sketch4

+++ zoomimage-core-glide/src/main/kotlin/com/github/panpf/zoomimage/glide/GlideTileBitmap.kt
@@ -35,7 +35,7 @@ internal class GlideTileBitmap(
 
     override val height: Int = bitmap.height
 
-    override val byteCount: Int = bitmap.byteCount
+    override val byteCount: Long = bitmap.byteCount.toLong()
 
     override val isRecycled: Boolean
         get() = bitmap.isRecycled
