+++ zoomimage-core/src/androidMain/kotlin/com/github/panpf/zoomimage/subsampling/AndroidTileBitmap.kt
@@ -38,7 +38,7 @@ private class AndroidTileBitmapImpl(
 
     override val height: Int = bitmap.height
 
-    override val byteCount: Int = bitmap.byteCount
+    override val byteCount: Long = bitmap.byteCount.toLong()
 
     override val isRecycled: Boolean
         get() = bitmap.isRecycled
